refactor(frontend): group rich-text rendering into markdown/
Move the Markdown renderer, citation tooltips and references, and the
rich-text accordions into a markdown/ folder and update importers. Add a
README marking the folder display-only. Retarget Markdown's stylelint
:global exemption to the new path.
