+++ frontend/.stylelintrc.cjs
@@ -135,7 +135,7 @@ module.exports = {
     // Components with deliberate :global usage (rendered HTML from
     // external sources). These are exceptions, not cleanup targets.
     {
-      files: ['src/lib/components/Prose.svelte', 'src/lib/components/Markdown.svelte'],
+      files: ['src/lib/components/Prose.svelte', 'src/lib/components/markdown/Markdown.svelte'],
       rules: {
         'selector-pseudo-class-disallowed-list': null,
         'selector-pseudo-class-no-unknown': null,
