+++ frontend/src/lib/components/markdown/README.md
@@ -1,0 +1,6 @@
+# markdown
+
+This folder contains rich text rendering components — markdown output plus the
+citation tooltips and references that get layered onto it.
+
+These components only do display; editing markdown and citations lives elsewhere.
