+++ frontend/src/lib/components/markdown/ReferencesSection.svelte
