+++ frontend/src/lib/components/markdown/citation-refs.ts
