+++ frontend/src/routes/cabinets/[slug]/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import client from '$lib/api/client';
   import AttributionLine from '$lib/components/provenance/AttributionLine.svelte';
-  import Markdown from '$lib/components/Markdown.svelte';
+  import Markdown from '$lib/components/markdown/Markdown.svelte';
   import MachineCard from '$lib/components/cards/MachineCard.svelte';
   import PaginatedSection from '$lib/components/grid/PaginatedSection.svelte';
   import { createPaginatedLoader } from '$lib/paginated-loader.svelte';
