+++ frontend/src/routes/franchises/[slug]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import AttributionLine from '$lib/components/provenance/AttributionLine.svelte';
-  import Markdown from '$lib/components/Markdown.svelte';
+  import Markdown from '$lib/components/markdown/Markdown.svelte';
   import TitleList from '$lib/components/TitleList.svelte';
 
   let { data } = $props();
