+++ frontend/src/routes/themes/[slug]/+page.svelte
@@ -4,7 +4,7 @@
   import HierarchicalTaxonomyChildrenAccordion from '$lib/components/HierarchicalTaxonomyChildrenAccordion.svelte';
   import HierarchicalTaxonomyMobileMetaBar from '$lib/components/HierarchicalTaxonomyMobileMetaBar.svelte';
   import MachineCard from '$lib/components/cards/MachineCard.svelte';
-  import Markdown from '$lib/components/Markdown.svelte';
+  import Markdown from '$lib/components/markdown/Markdown.svelte';
 
   let { data } = $props();
   let theme = $derived(data.profile);
