+++ frontend/src/routes/titles/[slug]/+page.svelte
@@ -6,9 +6,9 @@
   import MachineCard from '$lib/components/cards/MachineCard.svelte';
   import MediaGrid from '$lib/components/media/MediaGrid.svelte';
   import ModelSpecsSidebar from '$lib/components/ModelSpecsSidebar.svelte';
-  import RichTextOverviewAccordion from '$lib/components/RichTextOverviewAccordion.svelte';
-  import RichTextReferencesAccordion from '$lib/components/RichTextReferencesAccordion.svelte';
-  import { createRichTextAccordionState } from '$lib/components/rich-text-accordion-state.svelte';
+  import RichTextOverviewAccordion from '$lib/components/markdown/RichTextOverviewAccordion.svelte';
+  import RichTextReferencesAccordion from '$lib/components/markdown/RichTextReferencesAccordion.svelte';
+  import { createRichTextAccordionState } from '$lib/components/markdown/rich-text-accordion-state.svelte';
   import RelatedTitlesSection from '$lib/components/RelatedTitlesSection.svelte';
   import CreateFirstModelPrompt from '$lib/components/CreateFirstModelPrompt.svelte';
   import { titleAreaEditActionContext } from '$lib/components/editors/edit-action-context';
